+++ components/Form2/SimpleWorkoutDisplay.jsx
@@ -1,0 +1,20 @@
+export default function SimpleWorkoutDisplay({ data }) {
+  const workoutDisplayed = data.workouts.find(
+    (workout) => workout._id === "68834cb36f8968f2fa09c973"
+  );
+  /////////////////////////////////
+  return (
+    <>
+      <h2>{workoutDisplayed.name}</h2>
+      <ul>
+        {workoutDisplayed.exercises.map((exerciseInWorkout, index) => {
+          const idInWorkouts = exerciseInWorkout.exerciseId;
+          const exerciseInCollection = data.exercises.find(
+            (exerciseInCollection) => exerciseInCollection._id === idInWorkouts
+          );
+          return <li key={index}>{exerciseInCollection.name}</li>;
+        })}
+      </ul>
+    </>
+  );
+}
